Feature: settings button for saved searches notifications (#949)
* Use `UITableView.cellLayoutMarginsFollowReadableWidth`

To set the readable content guide relative margins. In this way we still
support the `readableContentGuide` but in a more reusable way as
suggested by Håkon in https://github.com/finn-no/FinniversKit/pull/943#issuecomment-688126245

* Add "more" button to the notification section header

This button will be used to present a bottom sheet for the save search
settings.

This commit includes 2 new requirements for the notification header

- `includeMoreButton` for the `NotificationCenterHeaderViewModel`
- `notificationCenterView(_:segment:didSelectMoreButtonIn)` delegate
method

This means this is a breaking change for FinnUI.

By default the "more" button will be hidden until the view model
indicates it needs it.